+++ dribdat/templates/public/projectpost.html
@@ -9,7 +9,7 @@
 {% block content %}
 
   {% if (suggestions or stage) and project.progress and project.progress < 50 %}
-    <div class="alert alert-secondary float-right resource-list project-tips profile-projects">
+    <div class="alert float-right resource-list project-tips profile-projects">
       <button type="button" class="close" data-dismiss="alert" aria-label="Dismiss" title="Hide">
         <span aria-hidden="true">&times;</span>
       </button>
